% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{get_immodata}
\alias{get_immodata}
\title{Websrcapping of Immoscout24.ch data by city}
\usage{
get_immodata(city_vector)
}
\arguments{
\item{city_vector}{A vector containing the name of the different cities}
}
\value{
A dataframe containing the number of rooms, m2, price, address, post code
and city of the different accomodations available.
}
\description{
This function enables to collect the data about the location market
from Immoscout24.ch. For given cities, on can retrieve the housings available for renting.
}
\examples{
 get_immodata(c("lugano", "lausanne"))
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
