% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{get_immodata2}
\alias{get_immodata2}
\title{Websrcapping of Immoscout24.ch data by city adding coordinates and predicted
price}
\usage{
get_immodata2(city_vector)
}
\arguments{
\item{city_vector}{A vector containing the name of the different cities}
}
\value{
A dataframe containing the number of rooms, m2, price, address,
city, longitude and latitude of the different accomodations available and their
estimated price. Ignores housings with prices less than 300 and higher than 10000
and containing missing data.
}
\description{
This function enables to collect the data about the location market
from Immoscout24.ch. For given cities, one can retrieve the housings available
for renting.
It adds the coordinates of the addresses and the predicted price of a good.
Note that this function requires an API code from Google and which is costly.
Therefore prefere to use the simple get_immodata function to avoid extraordinary costs.
}
\examples{
get_immodata2(c("bussigny"))
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
