% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{my_theme}
\alias{my_theme}
\title{Nice theme for ggplot graphs}
\usage{
my_theme(base_size = 10, base_family = "sans")
}
\arguments{
\item{base_size}{The base size, no need to change it.}

\item{base_family}{The base_family, no need to change it}
}
\description{
This function enables to add a beautiful theme to ggplot graphs
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
