% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{plot.pred}
\alias{plot.pred}
\title{Interactive plot of a "pred" object}
\usage{
\method{plot}{pred}(pred_object)
}
\arguments{
\item{pred_object}{An objet of class "pred", which was a dataframe inputted
in the predict_price function.}
}
\value{
A plot of the estimated prices against the real prices.
}
\description{
This function enables to plot a "pred" object and to retrieve
the estimated values compared to the real prices on the market. Thanks to a
shiny app, it is possible to retrieve the characteristic of the "points" by
clicking on them or selecting multiple at the same time.
}
\details{
One has to be careful using this function as the model to obtain the
predicted values has been built on the same dataset in which it forecasts the
price. Overfitting is clearly present and may alter the predictions.
}
\examples{
cities <- get_immodata(c("bussigny", "nyon"))
predictions <- predict_price(cities)
plot(predictions)
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
