% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{predict_price}
\alias{predict_price}
\title{Create a new "pred" object}
\usage{
predict_price(housings, rooms, m2, city, model = "rf", seed = 1)
}
\arguments{
\item{housings}{A dataframe in the form of the ones outputted by the get_immodata
function.}

\item{rooms}{The number of rooms for a single housing estimation}

\item{m2}{The number of m2 for a single housing estimation}

\item{city}{The city of a single housing estimation}

\item{model}{A model supported by the caret function for regression ("gam", "rf",
"nnet", "svmRadialCost", "rpart", ...)}

\item{seed}{The seed to use}
}
\value{
A "pred" object of the expected prices of all observations of the
dataframe or the single housing
}
\description{
This function enables to predict the prices of housings based on
their number of rooms, size in square meters, city and using a particular model.
The user can provide either a dataframe in the form of the ones outputted by the
get_immodata or providing only rooms, m2 and city of a single housing.
}
\examples{
cities <- get_immodata(c("bussigny", "nyon"))
predict_price(cities) # based on a dataframe
predict_price(rooms = 3, m2 = 59, city = "nyon") # for an unique housing
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
