% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{summary.pred}
\alias{summary.pred}
\title{Extracting the prices of a "pred" object}
\usage{
\method{summary}{pred}(pred_object)
}
\arguments{
\item{pred_object}{An objet of class "pred", which was a dataframe inputted
in the predict_price function.}
}
\value{
A dataframe with a new column of estimated prices.
}
\description{
This function enables to extract the prices of a "pred" object
and to add a new column of the expected prices to the original dataframe.
}
\details{
One has to be careful using this function as the model to obtain the
predicted values has been built on the same dataset in which it forecasts the
price. Overfitting is clearly present and may alter the predictions.
}
\examples{
cities <- get_immodata(c("bussigny", "nyon"))
predictions <- predict_price(cities)
summary(predictions)
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
