% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immodata.R
\name{swissmap}
\alias{swissmap}
\title{Shiny app of the swiss real estate market}
\usage{
swissmap(all_cities)
}
\arguments{
\item{all_cities}{A dataframe resulting from the get_immodata2 function,
containing the different housings, their prices, rooms, m2, city, addresses,
longitude, latitude and predicted_price.}
}
\description{
This function enables to add a beautiful theme to ggplot graphs
}
\examples{
my_cities <- get_immodata2("lausanne", "nyon", "geneve")
swissmap(my_cities)
}
\author{
Germano David

Lomazzi Vincent

Bron Luca

Raisin Edgar

Grandadam Patrik
}
